Code refactoring: made one orm_services, fix docker file, made changes in functions
