--- lambda_bonus_source/requirements.txt
@@ -1,3 +1,0 @@
-requests==2.28.1
-SQLAlchemy==1.4.39
-psycopg2-binary==2.9.3
