--- lambda_check_payment_day/requirements.txt
@@ -1,9 +1,0 @@
-certifi==2022.6.15
-charset-normalizer==2.1.0
-greenlet==1.1.2
-idna==3.3
-psycopg2-binary==2.9.3
-requests==2.28.1
-SQLAlchemy==1.4.39
-urllib3==1.26.10
-
