--- lambda_worker_source/requirements.txt
@@ -1,8 +1,0 @@
-SQLAlchemy==1.4.39
-SQLAlchemy-Utils==0.38.2
-slack-sdk==3.17.2
-python-decouple==3.6
-python-dotenv==0.20.0
-psycopg2-binary==2.9.3
-requests==2.27.1
-
